Commands: silence dumb -Wextra warning from GCC
This is a rather unhelpful warning indicating that the ternary operator return
types are mismatched, returning an integer and an enumeral type.  Since the
integeral type is shorter to type, cast the enumeral type to `int`.  Silences
the -Wextra warning from GCC.

llvm-svn: 258548

+++ lldb/source/Commands/CommandObjectTarget.cpp
@@ -4083,11 +4083,11 @@ public:
             case eLookupTypeAddress:
                 if (m_options.m_addr != LLDB_INVALID_ADDRESS)
                 {
-                    if (LookupAddressInModule (m_interpreter, 
-                                               result.GetOutputStream(), 
-                                               module, 
-                                               eSymbolContextEverything | (m_options.m_verbose ? eSymbolContextVariable : 0),
-                                               m_options.m_addr, 
+                    if (LookupAddressInModule (m_interpreter,
+                                               result.GetOutputStream(),
+                                               module,
+                                               eSymbolContextEverything | (m_options.m_verbose ? static_cast<int>(eSymbolContextVariable) : 0),
+                                               m_options.m_addr,
                                                m_options.m_offset,
                                                m_options.m_verbose))
                     {
